refactor: extract into `IdentifyNameType`
So we can get rid of the `raise UnparsableName` and `rescue`s.
